file restructuring scipts / devnet
